+++ plugins/techdocs/src/reader/transformers/index.ts
@@ -26,4 +26,5 @@ export * from './removeMkdocsHeader';
 export * from './simplifyMkdocsFooter';
 export * from './onCssReady';
 export * from './scrollIntoAnchor';
+export * from './scrollIntoNavigation';
 export * from './transformer';