+++ plugins/scaffolder-backend/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/plugin-scaffolder-backend
 
+## 1.5.0-next.1
+
+### Minor Changes
+
+- c4b452e16a: Starting the implementation of the Wizard page for the `next` scaffolder plugin
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.15.0-next.1
+  - @backstage/integration@1.3.0-next.1
+  - @backstage/plugin-catalog-backend@1.3.1-next.1
+
 ## 1.5.0-next.0
 
 ### Minor Changes