+++ plugins/techdocs-module-addons-contrib/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-techdocs-module-addons-contrib
 
+## 1.0.3-next.1
+
+### Patch Changes
+
+- ad35364e97: feat(techdocs): add edit button support for bitbucketServer
+- Updated dependencies
+  - @backstage/core-components@0.10.1-next.1
+  - @backstage/integration@1.3.0-next.1
+  - @backstage/plugin-techdocs-react@1.0.3-next.1
+
 ## 1.0.3-next.0
 
 ### Patch Changes